+++ flecsi/exec/leg/future.hh
@@ -34,12 +34,22 @@ struct future<Return, exec::launch_type_t::index> {
     legion_future_.wait_all_results(silence_warnings);
   } // wait
 
-  Return get(Color index = 0, bool silence_warnings = false) {
+  [[deprecated("pass to a task or use all")]] Return get(Color index = 0,
+    bool silence_warnings = false) {
     if constexpr(std::is_same_v<Return, void>)
       return legion_future_.get_void_result(index, silence_warnings);
     else
       return legion_future_.get_result<Return>(index, silence_warnings);
   } // get
+  template<class R = Return, class = std::enable_if_t<!std::is_void_v<R>>>
+  auto all() {
+    const Color n = size();
+    std::vector<R> ret;
+    ret.reserve(n);
+    for(Color i = 0; i < n; ++i)
+      ret.push_back(legion_future_.get_result<R>(i));
+    return ret;
+  }
 
   Color size() const {
     return legion_future_.get_future_map_domain().get_volume();