+++ tutorial/standalone/poisson/app/analyze.cc
@@ -15,7 +15,7 @@ poisson::action::analyze(control_policy & cp) {
   util::annotation::rguard<analyze_region> guard;
   double sum =
     reduce<task::diff, exec::fold::sum>(cp.m, ud(cp.m), sd(cp.m)).get();
-  sum = execute<task::scale>(cp.m, sum).get();
+  sum = reduce<task::scale, exec::fold::max>(cp.m, sum).get();
   const double l2 = sqrt(sum);
   flog(info) << "l2 error: " << l2 << std::endl;
 } // analyze